feat: 添加主题系统支持并优化UI组件
refactor(订单变更): 简化订单变更检测错误处理
feat(主题): 新增bevy_theme插件支持主题管理
feat(UI): 为滚动条和交易卡片添加主题支持
feat(音频): 添加通知音效资源
docs: 新增README和架构文档
style: 清理无用代码和简化实现

+++ src/weav3r/profit.rs
@@ -219,7 +219,7 @@ impl FavoritesData {
     /// 按利润排序，前 sec 秒的利润排序，然后是老的利润排序
     fn sort_profit(params: SortProfitParams, items: Vec<ProfitUserInfo>) -> Vec<ProfitUserInfo> {
         let now = crate::tools::time::get_current_time();
-        let recent_sec = now - params.recent_sec;
+        let recent_sec = now - params.recent_sec as u64;
         let mut recent_items: Vec<ProfitUserInfo> = items
             .clone()
             .into_iter()
@@ -517,7 +517,7 @@ pub struct FilterItem {
 
 #[derive(Debug, Clone, Default)]
 pub struct SortProfitParams {
-    pub recent_sec: u64,
+    pub recent_sec: u32,
 }
 
 impl ContentHashable for ProfitInfo {
